+++ segno_mimos/__init__.py
@@ -6,4 +6,4 @@
 # License: BSD License
 #
 
-__version__ = '0.1.4'
+__version__ = '0.1.5'