+++ CHANGES.rst
@@ -1,7 +1,16 @@
 Changes
 =======
 
-0.1.4 -- 2016-08-dd
+0.1.5 -- 2016-08-dd
+-------------------
+* Updated docs
+* Removed return statement from ``PyQRCode.png()``
+* Internal code changes
+* Renamed (internal) module ``segno_mimos.qrcode.img`` into ``_img`` to avoid
+  confusion with ``segno_mimos.qrcode.image``
+
+
+0.1.4 -- 2016-08-14
 -------------------
 * Initial release
 * Support for PyQRCode 1.2.1 and qrcode 5.3